Extract make_contract_class and compute_class_hash from code to separate module
